Move db packages to legacy.
