+++ politeiawww/legacy/cmsdatabase/cockroachdb/encoding.go
