+++ politeiawww/legacy/cmsuser.go
@@ -15,7 +15,7 @@ import (
 
 	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 )
 
 // processInviteNewUser creates a new user in the db if it doesn't already
