+++ politeiawww/legacy/totp.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/pquerna/otp"
 	"github.com/pquerna/otp/totp"
 )
