+++ politeiawww/legacy/user/localdb/encoding.go
