+++ politeiawww/sessions/sessions.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/google/uuid"
 	"github.com/gorilla/sessions"
 )
