+++ politeiawww/sessions/store.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/decred/politeia/politeiawww/user"
+	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/google/uuid"
 	"github.com/gorilla/securecookie"
 	"github.com/gorilla/sessions"
